Update CI + Refactor IO + Update gencon example with periodic bcs
